requirements.txt
anthropic>=0.34.0
python-dotenv>=1.0.0
streamlit>=1.37.0
pandas>=2.2.2
google-genai>=1.0.0
pydantic>=2.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
pillow>=10.0.0
aiofiles>=23.0.0
